requirements.txt
pymongo             # MongoDB driver
pinecone            # Pinecone Python client (updated package name)
sentence-transformers # For generating embeddings
python-dotenv       # To load environment variables from .env